Only send delete request upon save button
