+++ src/components/scenegraph/Entity.js
@@ -110,7 +110,6 @@ export default class Entity extends React.Component {
           {printEntity(entity, this.onDoubleClick)}
         </span>
         <span className="entityActions">
-          {cloneButton}
           {removeButton}
         </span>
       </div>
